Add symphony request debug information for site and super admins
