+++ app/views/requests/status.html.erb
@@ -9,4 +9,12 @@
   <hr />
   <%= render partial: 'searchworks_item_information' %>
   <%= render partial: 'shared/request_status_information' %>
+
+  <% if can? :debug, current_request %>
+    <hr />
+    <h4>Symphony Request</h4>
+    <pre class="well"><%= JSON.pretty_generate(current_request.symphony_request.request_params) %></pre>
+    <h4>Symphony Response</h4>
+    <pre class="well"><%= JSON.pretty_generate(current_request.symphony_response_data || {}) %></pre>
+  <% end %>
 </div>
